requirements.txt
Adafruit-Blinka==8.39.1
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==2.0.0
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-requests==3.2.8
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.62.0
Adafruit-PureIO==1.1.11
anyio==4.3.0
asgiref==3.8.1
board==1.0
certifi==2024.2.2
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorzero==2.0
Django==5.0.4
gpiozero==2.0.1
h11==0.14.0
httptools==0.6.1
idna==3.7
packaging==24.0
paho-mqtt==1.6.1
plotly==5.21.0
psycopg2==2.9.9
pyftdi==0.55.4
pyserial==3.5
python-dotenv==1.0.1
pyusb==1.2.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.5.1
sniffio==1.3.1
sqlparse==0.5.0
tenacity==8.2.3
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
